requirements.txt
bcrypt==3.2.2
cffi==1.15.1
click==8.1.3
dnspython==2.2.1
Flask==2.1.2
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
pycparser==2.21
pymongo==4.1.1
python-dateutil==2.8.2
six==1.16.0
Werkzeug==2.1.2
XlsxWriter==3.0.3
zipp==3.8.0
